Add commands for querying BigQuery through neoterm
This adds two useful keybinds; one to run the contents of the current
buffer as a SQL query in BigQuery, using the bq command line tool,
and one to do the same but with the --dry_run flag specified.

The --dry_run flag means that no data will be processed, so is a
useful way to do syntax and cost checks. It might be useful to bind
this to file save, but want to experiment with it first.