Add vim-signature
This is primarily to show an indicator for where marks exist in the
file. Works out of the box with no extra config needed. There's
a potential issue where I'm now starting to put a lot of overlapping
things in the signcolumn (this, gitgutter, ale), but it should be fine.